refactor: 유네임을 통한 유저 탈퇴 검증 추가

+++ backend/src/main/java/com/example/backend/user/service/UserService.java
@@ -1,6 +1,7 @@
 package com.example.backend.user.service;
 
 import com.example.backend.common.enums.ExceptionStatus;
+import com.example.backend.common.exceptions.BadRequestException;
 import com.example.backend.common.exceptions.NotFoundException;
 import com.example.backend.github.domain.GithubRepository;
 import com.example.backend.user.domain.User;
@@ -29,12 +30,19 @@ public class UserService {
   }
 
   @Transactional
-  public boolean deleteUser(Long id) {
-    return userRepository.findById(id)
-        .map(user -> {
-          userRepository.delete(user);
-          return true;
-        })
+  public void deleteUser(Long id, String inputUsername) {
+    User user = userRepository.findById(id)
         .orElseThrow(() -> new NotFoundException(ExceptionStatus.NOT_FOUND));
+
+    if (verifyUsername(user.getUsername(), inputUsername)) {
+      userRepository.delete(user);
+    } else {
+      throw new BadRequestException(ExceptionStatus.USERNAME_MISMATCH);
+    }
+  }
+
+  private boolean verifyUsername(String storedUsername, String inputUsername) {
+    return storedUsername.equals(inputUsername);
   }
 }
+
